--- a/.gitignore
+++ b/.gitignore
@@ -131,3 +131,6 @@ reports/
 # Serena
 .serena/cache/
 
+# Auto-generated API documentation (see Issue #92)
+docs/api-reference/
+