--- a/.kiro/specs/backup-tool-file-optimization/tasks.md
+++ b/.kiro/specs/backup-tool-file-optimization/tasks.md
@@ -1,6 +1,6 @@
 # Implementation Plan
 
-- [ ] 1. Backup file status verification and cleanup completion
+- [x] 1. Backup file status verification and cleanup completion
   - Verify that all backup files (*_old.js, *_original.js) have been properly removed from the project
   - Confirm that backup metadata exists in .cleanup-backups directory with proper file information
   - Update project documentation to reflect the completed cleanup status