--- a/.kiro/specs/backup-tool-file-optimization/tasks.md
+++ b/.kiro/specs/backup-tool-file-optimization/tasks.md
@@ -6,7 +6,7 @@
   - Update project documentation to reflect the completed cleanup status
   - _Requirements: 1.1, 1.2, 1.3, 1.4_
 
-- [ ] 2. API Documentation Generator analysis and preparation
+- [x] 2. API Documentation Generator analysis and preparation
   - Analyze the current structure and functionality of tools/api-doc-generator.js
   - Identify the main functional areas for splitting (parsing, generation, templating, validation)
   - Create the tools/api-doc-generator/ directory structure for split components