--- a/.kiro/specs/backup-tool-file-optimization/tasks.md
+++ b/.kiro/specs/backup-tool-file-optimization/tasks.md
@@ -13,7 +13,7 @@
   - Document the current API and interface requirements for backward compatibility
   - _Requirements: 2.1, 2.2, 5.2_
 
-- [ ] 3. Implement APIDocParser component
+- [x] 3. Implement APIDocParser component
   - Extract source code parsing logic from api-doc-generator.js into APIDocParser.js
   - Implement methods for extracting class definitions, method signatures, and comments
   - Add support for Japanese comment parsing and JSDoc annotation extraction