--- a/.kiro/specs/backup-tool-file-optimization/tasks.md
+++ b/.kiro/specs/backup-tool-file-optimization/tasks.md
@@ -21,7 +21,7 @@
   - Ensure the component is under 2,500 words for MCP tool compatibility
   - _Requirements: 2.1, 2.3, 4.1_
 
-- [ ] 4. Implement DocumentationGenerator component
+- [x] 4. Implement DocumentationGenerator component
   - Extract documentation generation logic into DocumentationGenerator.js
   - Implement methods for transforming parsed data into documentation structure
   - Add support for cross-reference generation and multilingual content handling