--- a/.kiro/specs/backup-tool-file-optimization/tasks.md
+++ b/.kiro/specs/backup-tool-file-optimization/tasks.md
@@ -29,7 +29,7 @@
   - Ensure the component is under 2,500 words for MCP tool compatibility
   - _Requirements: 2.1, 2.3, 4.1_
 
-- [ ] 5. Implement TemplateRenderer component
+- [x] 5. Implement TemplateRenderer component
   - Extract template rendering logic into TemplateRenderer.js
   - Implement methods for loading and processing Markdown templates
   - Add support for styling, formatting, and output file generation