… .kiro/specs/backup-tool-file-optimization/CHANGE_HISTORY.md
… .kiro/specs/backup-tool-file-optimization/design.md
… .kiro/specs/backup-tool-file-optimization/requirements.md
… CLAUDE.md
… src/core/LeaderboardManager_backup.js
… src/core/PWAManager_backup.js
… tests/analytics/ComparisonEngine.test.js
… tests/analytics/comparison-engine-tests/ANALYSIS.md
… tests/analytics/comparison-engine-tests/ComparisonEngineBasicTests.js
… tests/analytics/comparison-engine-tests/ComparisonEngineIntegrationTests.js
… tests/analytics/comparison-engine-tests/ComparisonEnginePerformanceTests.js
… tools/api-doc-generator.js
… tools/api-doc-generator/APIDocValidator.js
… tools/api-doc-generator/DocumentationGenerator.js
… tools/api-doc-generator/TemplateRenderer.js